small fixes and typos

+++ createGraph.cpp
@@ -38,12 +38,12 @@ int main(int argc, char **argv)
     float bitcoinAmount;
 
     // The four lines of code below taken from: https://stackoverflow.com/questions/3072795/how-to-count-lines-of-a-file-in-c (Billy ONeal's answer)
-    int linesNumber = 0;
+    long long linesNumber = 0;
     ifstream in(argv[2]);
     string unused;
     while (getline(in, unused))++linesNumber;
 
-    for(int i=0;i<linesNumber;i++)
+    for(long long i=0;i<linesNumber;i++)
     {
         fscanf(readTxEdgesTimes,"%i %i %f %i", &inputBitcoinAddress, &outputBitcoinAddress, &bitcoinAmount, &timeStamp);
         fprintf(saveUsersGraph, "%i %i %.0lf %i\n",userId[inputBitcoinAddress],userId[outputBitcoinAddress],bitcoinAmount,timeStamp);
